Add the shunts
The lesser of two evils in my opinion.  This is a poor man’s
arbitrary-precision arithmetic library.  Because the range of 64-bit
integers comes nowhere near the range of values a Double can represent,
we shunt all values in the log-adjusted range calculation into
floating-point land and calculate there before shunting back down to
integer land.  I fear we may simply be shifting the goalposts, but
until I have more robust test cases I can’t be sure this is the case.
For now, it works and I am sad.