add Gen.compose to allow imperative generator composition
Swift's lack of implicit conformance and polyvariadic functions makes
composing generators using map/zip or functional operators cumbersome
and/or infeasible for types with a large number of input parameters.

Not only is manually gathering many generated values to create a new
generator vulnerable to losing entropy, but it's also rife with
duplicated calls to manually split the StdGen. GenComposer fixes this by
encapsulating the logic of splitting the StdGen, as well as offering
generic functions to remove even more boilerplate for the default use
case (T.arbitrary).